2025.1_1/requirements-ubuntu-18.04.txt
pyDigitalWaveTools==1.2
tensorflow==2.15.0.post1
torch==2.5.1
numpy==1.26.4
matplotlib==3.8.3
scikit-learn==1.4.0
pandas==2.2.0
pytest==8.0.0
pytest-cov==4.1.0
pydot==2.0.0
graphviz==0.20.1
jupyter==1.0.0
seaborn==0.13.2
sympy==1.13.1
calmjs==3.4.4
tabulate==0.9.0
tf-keras==2.15.0
onnx==1.15.0
onnxruntime==1.19.2
qonnx @ git+https://github.com/fastmachinelearning/qonnx.git
QKeras @ git+https://github.com/jmitrevs/qkeras.git@qrecurrent_unstack
PySide6==6.7.2
docopt==0.6.2
pyaml==23.12.0
tensorboard-plugin-wit==1.8.1
yarg==0.1.9
zipp==3.17.0
backcall==0.2.0
pipreqs==0.5.0
importlib-metadata==7.0.1
importlib-resources==6.1.2
jax==0.4.16
